+++ oereb_client/static/js/service/map.js
@@ -9,14 +9,14 @@ goog.require('oereb');
  * @param {angular.$location} $location Angular location service.
  * @param {angular.$timeout} $timeout Angular $timeout service.
  * @param {string} oerebBaseLayerConfig The base layer configuration.
- * @param {string} oerebInitialExtentConfig The initial extent configuration.
+ * @param {string} oerebViewConfig The view configuration.
  * @param {string} oerebAvailabilityConfig The availability layer configuration.
  * @constructor
  * @ngInject
  * @ngdoc service
  * @ngname MapService
  */
-oereb.MapService = function($q, $http, $location, $timeout, oerebBaseLayerConfig, oerebInitialExtentConfig,
+oereb.MapService = function($q, $http, $location, $timeout, oerebBaseLayerConfig, oerebViewConfig,
                             oerebAvailabilityConfig) {
 
   this.$q_ = $q;
@@ -24,13 +24,13 @@ oereb.MapService = function($q, $http, $location, $timeout, oerebBaseLayerConfig
   this.$location_ = $location;
   this.$timeout_ = $timeout;
   this.baseLayerConfig_ = angular.fromJson(oerebBaseLayerConfig);
-  this.initialExtentConfig_ = angular.fromJson(oerebInitialExtentConfig);
+  this.viewConfig_ = angular.fromJson(oerebViewConfig);
   this.availabilityConfig_ = angular.fromJson(oerebAvailabilityConfig);
 
   // Initial extent
-  var mapX = parseFloat(this.$location_.search()['map_x']) || this.initialExtentConfig_['map_x'];
-  var mapY = parseFloat(this.$location_.search()['map_y']) || this.initialExtentConfig_['map_y'];
-  var mapZoom = parseInt(this.$location_.search()['map_zoom']) || this.initialExtentConfig_['map_zoom'];
+  var mapX = parseFloat(this.$location_.search()['map_x']) || this.viewConfig_['map_x'];
+  var mapY = parseFloat(this.$location_.search()['map_y']) || this.viewConfig_['map_y'];
+  var mapZoom = parseInt(this.$location_.search()['map_zoom']) || this.viewConfig_['map_zoom'];
 
   // Define LV03 projection
   proj4.defs(
@@ -80,7 +80,8 @@ oereb.MapService = function($q, $http, $location, $timeout, oerebBaseLayerConfig
   var view = new ol.View({
     projection: this.proj_,
     center: [mapX, mapY],
-    zoom: mapZoom
+    zoom: mapZoom,
+    resolutions: this.viewConfig_['resolutions']
   });
   view.setMinZoom(5);
   view.setMaxZoom(18);